Use `create()` factory method on paste
The constructor expects a duration value as its only parameter, and not
a `MediaItem`.  The text based paste handler, as part of attempting to
detect which format is being pasted, ends up creating a `MediaItem`, and
attempts to convert it to specifically a `ChaptersJson`, passing it in
as though it had a proper copy constructor.

In order to properly convert one `MediaItem` to another, either the
`from()` method should be used to import one `MediaItem` into an
existing one, or the `create()` should be used, which first creates a
new `MediaItem` then calls `from()` to import the data into the new
`MediaItem`.